requirements.txt
aniso8601==3.0.2
click==6.7
Flask==1.0.2
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
mistune==0.8.3
passlib==1.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
PyYAML==3.13
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.23
uWSGI==2.0.17.1
Werkzeug==0.14.1
